+1 player logic
