+++ functions/onePlayerFuncs.js
@@ -1,0 +1,3 @@
+export function singlePlayerGameInit(playgrid) {
+
+}
