+++ index.css
@@ -18,7 +18,7 @@ body::after {
     width: 100vw;
     top: 0;
     right: 0;
-    background: url(./mainBack.png);
+    background: url(./icon-imgs/mainBack.png);
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
@@ -34,7 +34,7 @@ body {
 
 .main {
     display: flex;
-    background: linear-gradient(rgb(145 205 165), rgb(140 158 169));
+    background: linear-gradient(rgb(181 228 197), rgb(140 158 169));
     justify-content: center;
     flex-direction: column;
     text-align: center;
@@ -52,6 +52,14 @@ body {
 
 .main>h4 {
     opacity: 0.2;
+}
+
+#winCounter {
+    display: flex;
+    width: 50%;
+    margin-inline: auto;
+    font-size: x-large;
+    justify-content: space-around;
 }
 
 #turnDisplay {
@@ -87,6 +95,9 @@ body {
     grid-template-columns: 33% 33% 33%;
 }
 
+.box:hover {
+    background: #000000c9;
+}
 .box {
     cursor: pointer;
     box-shadow: 1px 2px 5px rgba(0, 0, 0, 0.143), 0px -2px 10px rgba(0, 0, 0, 0.1);
@@ -95,7 +106,7 @@ body {
     margin: 15%;
     /* margin-top: 2px; */
     border-radius: 12px;
-    background: #ffffff5d;
+    background: #000000e0;
     pointer-events: all;
 }
 
@@ -146,15 +157,12 @@ body {
     opacity: 1;
     width: 5em;
     margin-top: 5px;
-    background: rgba(234, 233, 238, 0.71);
+    /* background: rgba(234, 233, 238, 0.71); */
     padding: 0.3em 0.6em;
     border-radius: 0.5em;
     color: rgba(18, 18, 18, 0.73);
 }
 
-#DP {
-    background: #59b8e8;
-}
 #reset {
     text-align: center;
     font-size: large;
