+++ index.html
@@ -7,13 +7,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./index.css">
     <title>TicTacToe</title>
-    <link rel="icon" type="image/x-icon" href="./favicon.png">
+    <link rel="icon" type="image/x-icon" href="./icon-imgs/favicon.png">
 </head>
         
 <body>
     <div class="main" id="main">
         <h1 id="website-title">TicTacToe</h1>
-        <p id="message">BTW Your Opponent is Dumb</p>
         <span id="PlayMode">
             <button id="SP">
                 <img src="https://cdn-icons-png.flaticon.com/128/2354/2354573.png" alt="">
@@ -28,16 +27,20 @@
             <p id="Player-1-text">Player 1</p>
             <p id="Player-2-text">Player 2</p>
         </span>
+        <span id="winCounter">
+            <p id="Player-1-wins">0</p>
+            <p id="Player-2-wins">0</p>  
+        </span>
         <div id="main-box">
-            <div id="11" class="box" data-id="11"></div>
-            <div id="12" class="box" data-id="12"></div>
-            <div id="13" class="box" data-id="13"></div>
-            <div id="21" class="box" data-id="21"></div>
-            <div id="22" class="box" data-id="22"></div>
-            <div id="23" class="box" data-id="23"></div>
-            <div id="31" class="box" data-id="31"></div>
-            <div id="32" class="box" data-id="32"></div>
-            <div id="33" class="box" data-id="33"></div>
+            <div id="11" class="box" data-id="1"></div>
+            <div id="12" class="box" data-id="2"></div>
+            <div id="13" class="box" data-id="3"></div>
+            <div id="21" class="box" data-id="4"></div>
+            <div id="22" class="box" data-id="5"></div>
+            <div id="23" class="box" data-id="6"></div>
+            <div id="31" class="box" data-id="7"></div>
+            <div id="32" class="box" data-id="8"></div>
+            <div id="33" class="box" data-id="9"></div>
         </div>
         <div id="reset">
             <p>new game starts in </p>
@@ -45,6 +48,6 @@
         </div>
     </div>
 </body>
-<script src="./index.js" type="text/javascript"></script>
+<script src="./index.js" type="module"></script>
 
 </html>
